+++ src/lib/list/list-super-partials/index.ts
@@ -1,0 +1,5 @@
+import listSuperReducer from './reducer';
+import getInitialState from './state';
+export * from './type';
+
+export { listSuperReducer, getInitialState };
