+++ src/lib/list/ui/loader.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Loader = (): JSX.Element => <p>Loading...</p>;
+
+export default Loader;
